ovl: fix open in stacked overlay [ Upstream commit 1c8a47df36d72ace8cf78eb6c228aa0f8027d3c2 ]
If two overlayfs filesystems are stacked on top of each other, then we
need
recursion in ovl_d_select_inode().

I guess d_backing_inode() is supposed to do that.  But currently it
doesn't
and that functionality is open coded in vfs_open().  This is now copied
into ovl_d_select_inode() to fix this regression.

Reported-by: default avatarAlban Crequy <alban.crequy@gmail.com>
Signed-off-by: default avatarMiklos Szeredi <miklos@szeredi.hu>
Fixes: 4bacc9c9234c ("overlayfs: Make f_path always point to the
overlay...")
Cc: David Howells <dhowells@redhat.com>
Cc: <stable@vger.kernel.org> # v4.2+
Signed-off-by: default avatarSasha Levin <sashal@kernel.org>