ovl: initialize error in ovl_copy_xattr commit 520da69d265a91c6536c63851cbb8a53946974f0 upstream.
In ovl_copy_xattr, if all the xattrs to be copied are overlayfs private
xattrs, the copy loop will terminate without assigning anything to the
error variable, thus returning an uninitialized value.

If ovl_copy_xattr is called from ovl_clear_empty, this uninitialized
error
value is put into a pointer by ERR_PTR(), causing potential invalid
memory
accesses down the line.

This commit initialize error with 0. This is the correct value because
when
there's no xattr to copy, because all xattrs are private, ovl_copy_xattr
should succeed.

This bug is discovered with the help of INIT_STACK_ALL and clang.

Signed-off-by: default avatarYuxuan Shui <yshuiv7@gmail.com>
Link: https://bugs.chromium.org/p/chromium/issues/detail?id=1050405

Fixes: 0956254a2d5b ("ovl: don't copy up opaqueness")
Cc: stable@vger.kernel.org # v4.8
Signed-off-by: default avatarAlexander Potapenko <glider@google.com>
Signed-off-by: Miklos Szeredi's avatarMiklos Szeredi
<mszeredi@redhat.com>
Signed-off-by: default avatarGreg Kroah-Hartman
<gregkh@linuxfoundation.org>
Signed-off-by: default avatarLee Jones <lee.jones@linaro.org>
Change-Id: I4fbcfd04bf39bad0756733bd85cef3de990090f3