ovl: Fix info leak in ovl_lookup_temp() commit 6a45b3628ce4dcf7498b39c87d475bab6e2a9b24 upstream.
The function uses the memory address of a struct dentry as unique id.
While the address-based directory entry is only visible to root it is
IMHO
still worth fixing since the temporary name does not have to be a kernel
address.  It can be any unique number.  Replace it by an atomic integer
which is allowed to wrap around.

Signed-off-by: default avatarRichard Weinberger <richard@nod.at>
Reviewed-by: default avatarKees Cook <keescook@chromium.org>
Signed-off-by: Miklos Szeredi's avatarMiklos Szeredi
<mszeredi@redhat.com>
Fixes: e9be9d5e

 ("overlay filesystem")
 Signed-off-by: default avatarGreg Kroah-Hartman
 <gregkh@linuxfoundation.org>
 Signed-off-by: default avatarLee Jones <lee.jones@linaro.org>
 Change-Id: I6f762559186098e73ad6ea14b85b759b87cd8778