ovl: fix failure to fsync lower dir commit d796e77f1dd541fe34481af2eee6454688d13982 upstream.
As a writable mount, it is not expected for overlayfs to return
EINVAL/EROFS for fsync, even if dir/file is not changed.

This commit fixes the case of fsync of directory, which is easier to
address, because overlayfs already implements fsync file operation for
directories.

The problem reported by Raphael is that new PostgreSQL 10.0 with a
database in overlayfs where lower layer in squashfs fails to start.
The failure is due to fsync error, when PostgreSQL does fsync on all
existing db directories on startup and a specific directory exists
lower layer with no changes.

Reported-by: default avatarRaphael Hertzog <raphael@ouaza.com>
Signed-off-by: default avatarAmir Goldstein <amir73il@gmail.com>
Tested-by: default avatarRaphaël Hertzog <hertzog@debian.org>
Signed-off-by: Miklos Szeredi's avatarMiklos Szeredi
<mszeredi@redhat.com>
Signed-off-by: default avatarGreg Kroah-Hartman
<gregkh@linuxfoundation.org>
Signed-off-by: default avatarLee Jones <lee.jones@linaro.org>
Change-Id: I8f26ab1e2360d8522dc4ffeb5b0015ba6f5d282f